backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
google-genai
python-dotenv
pydantic
